Refactor: Improve ViewModel interactions, repository structure, and Proguard configuration.
This commit addresses several points from the issue:

1.  **ViewModel Interaction (Issue 2):**
    I refactored `PokemonListScreen` and `PokemonDetailScreen` to use lambda functions for event callbacks to child composables (`EmptyErrorState`, `PokemonList`, `PokemonDetailed`, etc.). This replaces direct ViewModel passing, improving composable reusability and testability.

2.  **Repository Location (Issue 4):**
    I moved the `PokemonRepo.kt` (repository interface) from the `repo` package to the `domain` package within the `core-base` module. All import statements have been updated accordingly. This aligns the project with clean architecture principles.

3.  **Unused Files (Issue 3):**
    I removed empty `proguard-rules.pro` and `consumer-rules.pro` files from `core-base`, `core-ui`, `features/pokemonDetail`, and `features/pokemonList` modules.
    I also updated the corresponding `build.gradle.kts` files to remove references to these deleted files, simplifying the build configurations.

4.  **LazyColumn Performance (Issue 1):**
    I reviewed the usage of `LazyColumn` extensions in `PokemonListScreen`. The existing pagination logic, which relies on `LazyListState` attributes, was confirmed to be correctly using `derivedStateOf`. No further changes were needed for this point based on the initial issue description.
