+++ core-base/src/main/java/com/diegaspar/core_base/di/dataModule.kt
@@ -5,7 +5,7 @@ import com.diegaspar.core_base.data.persistence.PokemonLocalDataSourceImpl
 import com.diegaspar.core_base.data.persistence.mapper.PokemonEntityMapper
 import com.diegaspar.core_base.data.remote.PokemonRemoteDataSource
 import com.diegaspar.core_base.data.remote.PokemonRemoteDataSourceImpl
-import com.diegaspar.core_base.repo.PokemonRepo
+import com.diegaspar.core_base.domain.PokemonRepo
 import com.diegaspar.core_base.repo.PokemonRepoImpl
 import org.koin.dsl.module
