+++ features/pokemonList/src/main/java/com/diegaspar/pokemonlist/domain/GetPokemonsUseCase.kt
@@ -1,7 +1,7 @@
 package com.diegaspar.pokemonlist.domain
 
 import com.diegaspar.core_base.domain.PokemonListPaginated
-import com.diegaspar.core_base.repo.PokemonRepo
+import com.diegaspar.core_base.domain.PokemonRepo
 
 class GetPokemonsUseCase(private val pokemonRepository: PokemonRepo) {
     suspend operator fun invoke(params: Params): PokemonListPaginated =
